Conflicts:
	everware/spawner.py

+++ build_tools/test_frontend.sh
@@ -9,18 +9,28 @@ NPROC=2
 
 echo "In" `pwd`
 
-# Start a hub that our tests can interact with
-echo "Starting everware"
+OPTS="-f build_tools/frontend_test_config.py --no-ssl --debug $1"
 
-OPTS="-f build_tools/frontend_test_config.py --no-ssl --debug"
+# Start a hub that our tests can interact with
+echo "Starting everware-server($OPTS)"
 everware-server ${OPTS} > $LOG 2>&1 &
 HUB_PID=$!
 sleep 3
 
+if [[ `pgrep -f everware-server` == "" ]] ; then 
+    echo "Error starting"
+    tail $LOG 
+    exit 1
+fi
+
 echo "Start running frontend tests"
-[ -z "$UPLOADDIR" ] && echo "no UPLOADDIR defined" && exit 1
+if [ -z "$UPLOADDIR" ] ; then 
+	echo "no UPLOADDIR defined" 
+	kill ${HUB_PID}
+	exit 1
+fi
 [ -d $UPLOADDIR ] && rm -rf $UPLOADDIR/*
-nose2 -v --start-dir frontend_tests || FAIL=1
+nose2 -v -N $NPROC --start-dir=frontend_tests || FAIL=1
 
 if [ -f $LOG ]; then
     echo ">>> Frontend test hub log:"
@@ -28,6 +38,10 @@ if [ -f $LOG ]; then
     echo "<<< Frontend test hub log:"
     docker ps -a
 fi
+
+echo ">>> Frontend test client log"
+find $UPLOADDIR -name "*.log" | xargs cat
+echo "<<< Frontend test client log"
 
 kill ${HUB_PID}
 exit $FAIL
